(schematics) add configuration rule
