+++ projects/lib/schematics/make-static/utils/html-transformer.ts
