+++ projects/lib/tsconfig.lib.json
@@ -9,7 +9,7 @@
     "types": [],
     "lib": [
       "dom",
-      "es2018"
+      "es2022"
     ]
   },
   "exclude": [
