+++ projects/lib/tsconfig.schematics.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "baseUrl": ".",
     "lib": [
-      "es2020",
+      "es2022",
       "dom"
     ],
     "declaration": true,
@@ -21,7 +21,7 @@
     "skipLibCheck": true,
     "sourceMap": true,
     "strictNullChecks": true,
-    "target": "es6",
+    "target": "ES2022",
     "types": [
       "jasmine",
       "node"
